+++ codei/app/Config/Services.php
@@ -19,14 +19,12 @@ use CodeIgniter\Config\BaseService;
  */
 class Services extends BaseService
 {
-    /*
-     * public static function example($getShared = true)
-     * {
-     *     if ($getShared) {
-     *         return static::getSharedInstance('example');
-     *     }
-     *
-     *     return new \CodeIgniter\Example();
-     * }
-     */
+    public static function databaseCapabilityInspector(bool $getShared = true): \App\Services\DatabaseCapabilityInspector
+    {
+        if ($getShared) {
+            return static::getSharedInstance('databaseCapabilityInspector');
+        }
+
+        return new \App\Services\DatabaseCapabilityInspector();
+    }
 }